+++ packages/common/src/paths/index.ts
@@ -17,6 +17,7 @@ import * as path from 'path';
 
 export const USER_HOME = os.homedir();
 export const NITRIC_HOME = process.env.NITRIC_HOME || path.join(USER_HOME, '.nitric');
+export const TMP_DIR = path.join(NITRIC_HOME, 'tmp');
 export const STAGING_DIR = path.join(NITRIC_HOME, 'staging');
 export const STAGING_API_DIR = path.join(STAGING_DIR, 'apis');
 export const TEMPLATE_DIR = path.join(NITRIC_HOME, 'templates');
