+++ packages/plugins/aws/src/commands/deploy/aws.ts
@@ -14,7 +14,7 @@
 
 import { flags } from '@oclif/command';
 import { Deploy, DEPLOY_TASK_KEY, DeployResult } from '../../tasks/deploy';
-import { BaseCommand, wrapTaskForListr, Stack, constants } from '@nitric/cli-common';
+import { BaseCommand, wrapTaskForListr, Stack, constants, createBuildListrTask } from '@nitric/cli-common';
 import { Listr } from 'listr2';
 import path from 'path';
 import AWS from 'aws-sdk';
@@ -98,7 +98,7 @@ export default class AwsDeploy extends BaseCommand {
 		const stack = await Stack.fromFile(stackDefinitionPath);
 
 		const results = await new Listr<any>(
-			[wrapTaskForListr(new Deploy({ stack, account: accountId, region }))],
+			[createBuildListrTask(stack, 'aws'), wrapTaskForListr(new Deploy({ stack, account: accountId, region }))],
 			constants.DEFAULT_LISTR_OPTIONS,
 		).run();
